+++ Changes.md
@@ -76,6 +76,19 @@ Build
   - OpenColorIO : Updated to version 2.1.1.
   - Cortex : Updated to version 10.4.0.0.
 
+0.61.x.x (relative to 0.61.5.0)
+========
+
+Fixes
+-----
+
+- Spreadsheet : Fixed attempts to edit non-editable plugs when double-clicking on a boolean cell.
+
+API
+---
+
+- SpreadsheetUI : Added `addColumnButtonMenuSignal()`, which can be used to customise the column creation menu for spreadsheets.
+
 0.61.5.0 (relative to 0.61.4.0)
 ========